Text.CSL.Eval

Description

The CSL implementation

Documentation

evalLayout :: Layout -> EvalMode -> Bool -> [Locale] -> [MacroMap] -> [Option] -> Abbreviations -> Reference -> [Output]

Produce the output with a Layout, the EvalMode, a Bool True if the evaluation happens for disambiguation purposes, the Locale, the MacroMap, the position of the cite and the Reference.

evalSorting :: EvalMode -> [Locale] -> [MacroMap] -> [Option] -> [Sort] -> Abbreviations -> Reference -> [Sorting]

module Text.CSL.Eval.Common

module Text.CSL.Eval.Output