Text.CSL.Input.Bibutils

Description

Documentation

readBiblioFile :: FilePath -> IO [Reference]

Read a file with a bibliographic database. The database format is recognized by the file extension.

Supported formats are: json, mods, bibtex, biblatex, ris, endnote, endnotexml, isi, medline, and copac.

readBiblioString :: BibFormat -> String -> IO [Reference]

data BibFormat

Constructors

Json

Yaml

Bibtex

BibLatex

Ris

Endnote

EndnotXml

Isi

Medline

Copac

Mods

convertRefs :: Maybe MetaValue -> Either String [Reference]