Text.CSL.Pandoc

Documentation

processCites :: Style -> [Reference] -> Pandoc -> Pandoc

Process a Pandoc document by adding citations formatted according to a CSL style. Add a bibliography (if one is called for) at the end of the document.

processCites' :: Pandoc -> IO Pandoc

Process a Pandoc document by adding citations formatted according to a CSL style. The style filename is derived from the csl field of the metadata, and the references are taken from the references field or read from a file in the bibliography field.