Text.CSL.Proc

Description

This module provides functions for processing the evaluated Output for disambiguation and citation collapsing.

Documentation

data ProcOpts

Constructors

ProcOpts

Fields

bibOpts :: BibOpts

data BibOpts

Constructors

Select [(String, String)] [(String, String)]

Include [(String, String)] [(String, String)]

Exclude [(String, String)] [(String, String)]

newtype FieldVal

Constructors

FieldVal

Fields

unFieldVal :: (String, String)

procOpts :: ProcOpts

processCitations :: ProcOpts -> Style -> [Reference] -> Citations -> [Formatted]

With a Style, a list of References and the list of citation groups (the list of citations with their locator), produce the Formatted for each citation group.

processBibliography :: ProcOpts -> Style -> [Reference] -> [Formatted]

With a Style and the list of References produce the Formatted for the bibliography.

citeproc :: ProcOpts -> Style -> [Reference] -> Citations -> BiblioData

With a Style, a list of References and the list of Citations, produce the Formatted for each citation group and the bibliography.

procRefs :: Style -> [Reference] -> [Reference]

Given the CSL Style and the list of References sort the list according to the Style and assign the citation number to each Reference.

sortItems :: Show a => [(a, [Sorting])] -> [a]

procBiblio :: BibOpts -> Style -> [Reference] -> [[Output]]

With a Style and a sorted list of References produce the evaluated output for the bibliography.

subsequentAuthorSubstitute :: Bibliography -> [[Output]] -> [[Output]]

filterRefs :: BibOpts -> [Reference] -> [Reference]

procGroup :: Style -> [(Cite, Reference)] -> CitationGroup

Given the CSL Style and the list of Cites coupled with their References, generate a CitationGroup. The citations are sorted according to the Style.

formatBiblioLayout :: Formatting -> Delimiter -> [Output] -> [Output]

formatCitLayout :: Style -> CitationGroup -> Formatted

addAffixes :: Formatting -> [Output] -> [Output]

localModifiers :: Style -> Bool -> Cite -> Output -> Output

The Bool is True if we are formatting a textual citation (in pandoc terminology).

contribOnly :: Style -> Output -> Output