Text.CSL.Proc.Collapse

Description

This module provides functions for processing the evaluated Output for citation collapsing.

Documentation

collapseCitGroups :: Style -> [CitationGroup] -> [CitationGroup]

Collapse citations according to the style options.

getCollapseOptions :: Style -> [String]

Get the collapse option set in the Style for citations.

collapseNumber :: CitationGroup -> CitationGroup

groupCites :: [(Cite, Output)] -> [(Cite, Output)]

getYearAndSuf :: Output -> Output

collapseYear :: Style -> String -> CitationGroup -> CitationGroup

collapseYearSuf :: Bool -> String -> [(Cite, Output)] -> [Output]

collapseYearSufRanged :: [Output] -> [Output]

addCiteAffixes :: Cite -> Output -> Output

isNumStyle :: [Output] -> Bool

groupConsec :: [Int] -> [[Int]]

Group consecutive integers:

groupConsec [1,2,3,5,6,8,9] == [[1,2,3],[5,6],[8,9]]