Text.CSL.Style

Description

The Style types

Documentation

readCSLString :: String -> [Inline]

writeCSLString :: [Inline] -> String

newtype Formatted

Constructors

Formatted

Fields

unFormatted :: [Inline]

data Style

The representation of a parsed CSL style.

Constructors

Style

Fields

styleVersion :: String

styleClass :: String

styleInfo :: Maybe CSInfo

styleDefaultLocale :: String

styleLocale :: [Locale]

styleAbbrevs :: Abbreviations

csOptions :: [Option]

csMacros :: [MacroMap]

citation :: Citation

biblio :: Maybe Bibliography

data Locale

Constructors

Locale

Fields

localeVersion :: String

localeLang :: String

localeOptions :: [Option]

localeTerms :: [CslTerm]

localeDate :: [Element]

mergeLocales :: String -> Locale -> [Locale] -> [Locale]

With the defaultLocale, the locales-xx-XX.xml loaded file and the parsed Style cs:locale elements, produce the final Locale as the only element of a list, taking into account CSL locale prioritization.

data CslTerm

Constructors

CT

Fields

cslTerm :: String

termForm :: Form

termGender :: Gender

termGenderForm :: Gender

termSingular :: String

termPlural :: String

termMatch :: String

newTerm :: CslTerm

findTerm :: String -> Form -> [CslTerm] -> Maybe CslTerm

findTerm' :: String -> Form -> Gender -> [CslTerm] -> Maybe CslTerm

newtype Abbreviations

Constructors

Abbreviations

Fields

unAbbreviations :: Map String (Map String (Map String String))

type MacroMap = (String, [Element])

data Citation

Constructors

Citation

Fields

citOptions :: [Option]

citSort :: [Sort]

citLayout :: Layout

data Bibliography

Constructors

Bibliography

Fields

bibOptions :: [Option]

bibSort :: [Sort]

bibLayout :: Layout

type Option = (String, String)

mergeOptions :: [Option] -> [Option] -> [Option]

data Layout

Constructors

Layout

Fields

layFormat :: Formatting

layDelim :: Delimiter

elements :: [Element]

data Element

Constructors

Choose IfThen [IfThen] [Element]

Macro String Formatting

Const String Formatting

Variable [String] Form Formatting Delimiter

Term String Form Formatting Bool

Label String Form Formatting Plural

Number String NumericForm Formatting

Names [String] [Name] Formatting Delimiter [Element]

Substitute [Element]

Group Formatting Delimiter [Element]

Date [String] DateForm Formatting Delimiter [DatePart] String

data IfThen

Constructors

IfThen Condition Match [Element]

data Condition

Constructors

Condition

Fields

isType :: [String]

isSet :: [String]

isNumeric :: [String]

isUncertainDate :: [String]

isPosition :: [String]

disambiguation :: [String]

isLocator :: [String]

type Delimiter = String

data Match

Constructors

Any

All

None

match :: Match -> [Bool] -> Bool

data DatePart

Constructors

DatePart

Fields

dpName :: String

dpForm :: String

dpRangeDelim :: String

dpFormatting :: Formatting

defaultDate :: [DatePart]

data Sort

Constructors

SortVariable String Sorting

SortMacro String Sorting Int Int String

data Sorting

Constructors

Ascending String

Descending String

compare' :: String -> String -> Ordering

data Form

Constructors

Long

Short

Count

Verb

VerbShort

Symbol

NotSet

data Gender

Constructors

Feminine

Masculine

Neuter

data NumericForm

Constructors

Numeric

Ordinal

Roman

LongOrdinal

data DateForm

Constructors

TextDate

NumericDate

NoFormDate

data Plural

Constructors

Contextual

Always

Never

data Name

Constructors

Name Form Formatting NameAttrs Delimiter [NamePart]

NameLabel Form Formatting Plural

EtAl Formatting String

type NameAttrs = [(String, String)]

data NamePart

Constructors

NamePart String Formatting

isPlural :: Plural -> Int -> Bool

isName :: Name -> Bool

isNames :: Element -> Bool

hasEtAl :: [Name] -> Bool

data Formatting

Constructors

Formatting

Fields

prefix :: String

suffix :: String

fontFamily :: String

fontStyle :: String

fontVariant :: String

fontWeight :: String

textDecoration :: String

verticalAlign :: String

textCase :: String

display :: String

quotes :: Quote

stripPeriods :: Bool

noCase :: Bool

noDecor :: Bool

emptyFormatting :: Formatting

rmTitleCase :: Formatting -> Formatting

data Quote

Constructors

NativeQuote

ParsedQuote

NoQuote

unsetAffixes :: Formatting -> Formatting

mergeFM :: Formatting -> Formatting -> Formatting

data CSInfo

Constructors

CSInfo

Fields

csiTitle :: String

csiAuthor :: CSAuthor

csiCategories :: [CSCategory]

csiId :: String

csiUpdated :: String

data CSAuthor

Constructors

CSAuthor String String String

data CSCategory

Constructors

CSCategory String String String

data CiteprocError

Constructors

NoOutput

ReferenceNotFound String

data Output

The Output generated by the evaluation of a style. Must be further processed for disambiguation and collapsing.

Constructors

ONull

OSpace

OPan [Inline]

ODel String
A delimiter string.

OStr String Formatting
A simple String

OErr CiteprocError
Warning message

OLabel String Formatting
A label used for roles

ONum Int Formatting
A number (used to count contributors)

OCitNum Int Formatting
The citation number

ODate [Output]
A (possibly) ranged date

OYear String String Formatting
The year and the citeId

OYearSuf String String [Output] Formatting
The year suffix, the citeId and a holder for collision data

OName Agent [Output] [[Output]] Formatting
A (family) name with the list of given names.

OContrib String String [Output] [Output] [[Output]]
The citation key, the role (author, editor, etc.), the contributor(s), the output needed for year suf. disambiguation, and everything used for name disambiguation.

OLoc [Output] Formatting
The citation's locator

Output [Output] Formatting
Some nested Output

type Citations = [[Cite]]

data Cite

Constructors

Cite

Fields

citeId :: String

citePrefix :: Formatted

citeSuffix :: Formatted

citeLabel :: String

citeLocator :: String

citeNoteNumber :: String

citePosition :: String

nearNote :: Bool

authorInText :: Bool

suppressAuthor :: Bool

citeHash :: Int

emptyCite :: Cite

data CitationGroup

A citation group: the first list has a single member when the citation group starts with an "author-in-text" cite, the Formatting to be applied, the Delimiter between individual citations and the list of evaluated citations.

Constructors

CG [(Cite, Output)] Formatting Delimiter [(Cite, Output)]

data BiblioData

Constructors

BD

Fields

citations :: [Formatted]

bibliography :: [Formatted]

data CiteData

A record with all the data to produce the Formatted of a citation: the citation key, the part of the formatted citation that may be colliding with other citations, the form of the citation when a year suffix is used for disambiguation , the data to disambiguate it (all possible contributors and all possible given names), and, after processing, the disambiguated citation and its year, initially empty.

Constructors

CD

Fields

key :: String

collision :: [Output]

disambYS :: [Output]

disambData :: [[Output]]

disambed :: [Output]

sameAs :: [String]

citYear :: String

data NameData

Constructors

ND

Fields

nameKey :: Agent

nameCollision :: [Output]

nameDisambData :: [[Output]]

nameDataSolved :: [Output]

isPunctuationInQuote :: Style -> Bool

object' :: [Pair] -> Value

data Agent

Constructors

Agent

Fields

givenName :: [Formatted]

droppingPart :: Formatted

nonDroppingPart :: Formatted

familyName :: Formatted

nameSuffix :: Formatted

literal :: Formatted

commaSuffix :: Bool

parseNames :: Bool

emptyAgent :: Agent